squash! Remove/replace/refactor some utility functions
This commit changes the behavior of `lazyflow.roi.sliceToRoi`.

Out of range int indices were allowed and treated as if `i` was
`slice(i, i+1)`. Now such indexing produces an error, which is
consistent with the list/ndarray indexing.

    # Before:
    >>> sliceToRoi(42, (7, 8, 9))
    ([42, 0, 0], [43, 8, 9])
    # After:
    >>> sliceToRoi(42, (7, 8, 9))
    Traceback (most recent call last)
      ...
    ValueError

Indexing with negative values is now allowed, which is consistent with
the list/ndarray indexing.