Properly if object can be a dtype
also make sure not to break on anything that this check let's through by
instantiating the dtype rather than an object of a certain dtype.

Removes the following warning issued by numpy:
FutureWarning: Conversion of the second argument of issubdtype from `dtype` to
`np.generic` is deprecated. In future, it will be treated as
`np.object_ == np.dtype(dtype).type`.